requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
coverage==7.3.2
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.23
django-allauth==0.58.2
django-crispy-forms==1.14.0
django-jazzmin==2.6.0
django-social-share==2.3.0
django-summernote==0.8.20.0
django-widget-tweaks==1.5.0
gunicorn==21.2.0
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
ruff==0.1.11
sqlparse==0.4.4
